Removal of useEffect to load to Quill and various other updates for new RichTextEditor component
