+++ components/library/RichTextEditor/index.ts
@@ -2,5 +2,5 @@
 export { tabs } from "./tabs";
 export { Preview } from "./Preview/index";
 export const title = "Rich Text Editor";
-export const contributor = "Evan Forge"
-export const dependencies = ['react-quill','remix-utils/client-only'];
+export const contributor = "fabregas4"
+export const dependencies = ['react-quill'];
